0.3.0 release

+++ api-documentation/scripts/update-documentation.sh
@@ -1,4 +1,4 @@
-rm -r ./docs/n2e
+[ -d ./docs/n2e ] && rm -r ./docs/n2e
 cp ../openapi.json ./openapi/
 cp ../openapi.yml ./openapi/
 yarn docusaurus gen-api-docs all
